Fix: Demystified markdown logic (#1186)
* Fix: Demystified markdown logic

I have changes jsfiddle link with a similar jsfiddle. 

The difference between two jsfiddle is in javascript code, `marked()` has a comment mentioning its source as external library. 

This is important because it is easy to assume that `marked()` cmes from an external library instead of Vue, especially for a beginner.

* make dependencies apparent in markdown example